refactor(product_type): move views into dojo/product_type/ui/views.py [Phase 5]
Phase 5 of module reorg per AGENTS.md. Move dojo/product_type/views.py to
dojo/product_type/ui/views.py and update its two importers (dojo/organization/urls.py
and the counts unit test). product_type has no urls.py (routes live in
dojo/organization/urls.py), so only the views move.
